fix test and move base interruption to internals
