+++ chirho/dynamical/handlers/__init__.py
@@ -1,9 +1,8 @@
-from ..internals.solver import Solver  # noqa: F401
+from ..internals.solver import Interruption, Solver  # noqa: F401
 from .event_loop import InterruptionEventLoop  # noqa: F401
 from .interruption import (  # noqa: F401
     DynamicInterruption,
     DynamicIntervention,
-    Interruption,
     StaticBatchObservation,
     StaticInterruption,
     StaticIntervention,
